+++ src/components/tree/Tree.vue
@@ -22,7 +22,7 @@ import { ref, watch, PropType } from 'vue'
 import Model from "@/components/Model.vue";
 import TreeCore from "@/components/tree/TreeCore.vue";
 
-import { TreeNode } from "@/api/request";
+import type { TreeNode } from "@/api/request.type";
 
 const props = defineProps({
   isShow: {
@@ -43,7 +43,7 @@ let checked = ref<Array<String>>([])
 
 watch(
   () => [...props.codes],
-  (newValue, oldValue) => {
+  (newValue) => {
     checked.value = newValue
   }
 )
@@ -53,16 +53,13 @@ watch(
  */
 const track = (node: TreeNode) => {
   let isCheck = false;
-  if (checked.value.indexOf(node.code) === -1) {
+  if (checked.value.indexOf(node.code) == -1) {
     checked.value.push(node.code)
     isCheck = true
   } else {
     checked.value.splice(checked.value.indexOf(node.code), 1)
   }
-  let parents = recurrenceParents(node.code, props.datas)
-  if(isCheck){
-    checked.value.concat(parents)
-  }
+
   recurrenceChildren(node.children, isCheck)
 }
 /**
@@ -74,7 +71,7 @@ const recurrenceChildren = (children: Array<TreeNode>, isCheck: boolean) => {
   if (children && children.length > 0) {
     children.forEach((item: TreeNode) => {
       if (isCheck) {
-        if (checked.value.indexOf(item.code) === -1) {
+        if (checked.value.indexOf(item.code) == -1) {
           checked.value.push(item.code)
         }
       } else {
