+++ src/mock/assets/category.ts
@@ -1,7 +1,7 @@
 import { Random } from 'mockjs'
 
-import { Category } from '@/api/request'
-import { parse } from '@/api/util';
+import type { Category } from '@/api/request.type'
+import { parse } from '@/util';
 
 const datas: Array<Category> = [];
