+++ src/mock/assets/resource.ts
@@ -1,7 +1,7 @@
 import { Random } from 'mockjs'
 
-import { Resource } from '@/api/request'
-import { parse } from '@/api/util';
+import type { Resource } from '@/api/request.type'
+import { parse } from '@/util';
 
 const datas: Array<Resource> = [];
