+++ src/mock/system/authority.ts
@@ -1,7 +1,7 @@
 import { Random } from 'mockjs'
 
-import { Authority, Role } from '@/api/request'
-import { parse } from '@/api/util';
+import type { Authority, Role } from '@/api/request.type'
+import { parse } from '@/util';
 
 const datas: Array<Authority> = [
   {
