+++ src/mock/system/notification.ts
@@ -1,7 +1,7 @@
 import { Random } from 'mockjs'
 
-import { Notification } from '@/api/request'
-import { parse } from '@/api/util';
+import type { Notification } from '@/api/request.type'
+import { parse } from '@/util';
 
 const datas: Array<Notification> = [];
