+++ src/mock/system/region.ts
@@ -1,7 +1,7 @@
 import { Random } from 'mockjs'
 
-import { Region } from '@/api/request'
-import { parse } from '@/api/util';
+import type { Region } from '@/api/request.type'
+import { parse } from '@/util';
 
 const datas: Array<Region> = [];
