+++ src/views/Index.vue
@@ -17,7 +17,7 @@
 <script lang="ts" setup>
 import { ref, onMounted } from "vue";
 
-import { Authority } from "@/api/request";
+import type { Authority } from "@/api/request.type";
 
 import Drawer from "@/layouts/Drawer.vue";
 import Aside from "@/layouts/Aside.vue";
