+++ src/views/sign/SignUp.vue
@@ -55,7 +55,7 @@
 import { ref } from "vue";
 import Sign from "@/layouts/Sign.vue";
 
-import instance from "@/api";
+import { instance } from "@/api";
 
 let formData = ref({});
 const confirmPassword = ref("");
