+++ src/views/system/Group.vue
@@ -265,8 +265,8 @@ import Confirm from "@/components/Confirm.vue";
 import Model from "@/components/Model.vue";
 import Preview from "@/components/Preview.vue";
 
-import instance from "@/api";
-import { SERVER_URL, Group, Account } from "@/api/request";
+import { instance, SERVER_URL } from "@/api";
+import type { Group, Account } from "@/api/request.type";
 
 // 模态框参数
 let isEdit = ref(false);
