+++ src/views/system/Role.vue
@@ -159,8 +159,8 @@ import Confirm from "@/components/Confirm.vue";
 import Model from "@/components/Model.vue";
 import Tree from "@/components/tree/Tree.vue";
 
-import instance from "@/api";
-import { SERVER_URL, Role, Authority } from "@/api/request";
+import { instance, SERVER_URL } from "@/api";
+import type { Role, Authority } from "@/api/request.type";
 
 // 模态框参数
 let isEdit = ref(false);
